Basic outline of new input data structure
