+++ src/effortsharing/globalpathways.py
@@ -1,57 +1,10 @@
-# ======================================== #
-# Class that does the data reading
-# ======================================== #
-
-# =========================================================== #
-# PREAMBULE
-# Put in packages that we need
-# =========================================================== #
-
-import json
 from dataclasses import dataclass
 
 import numpy as np
 import pandas as pd
 import xarray as xr
 
-import effortsharing.regions as _regions
 from effortsharing.config import Config
-
-
-@dataclass
-class General:
-    countries: dict[str, str]
-    regions: dict[str, str]
-
-@dataclass
-class UNPopulation:
-    population: xr.Dataset
-    population_long: xr.Dataset
-
-@dataclass
-class JonesData:
-    xr_hist: xr.Dataset
-    xr_ghg_afolu: xr.Dataset
-    xr_ghg_agri: xr.DataArray
-    xr_edgar: xr.Dataset
-    xr_primap: xr.Dataset
-
-@dataclass
-class AR6Data:
-    xr_ar6_prevet: xr.Dataset
-    xr_ar6: xr.Dataset
-    ms_immediate: np.ndarray
-    ms_delayed: np.ndarray
-    xr_ar6_landuse: xr.Dataset
-    xr_ar6_C: xr.Dataset
-    xr_ar6_C_bunkers: xr.Dataset
-
-
-@dataclass
-class NonCO2Data:
-    xr_temperatures: xr.Dataset
-    xr_nonco2warmings: xr.Dataset
-    xr_nonco2warming_wrt_start: xr.Dataset
 
 
 # TODO: perhaps combine this class with the one before?
@@ -77,852 +30,6 @@ class GlobalCO2:
     xr_traj_ghg_excl: xr.Dataset
     xr_traj_co2_excl: xr.Dataset
     all_projected_gases: xr.Dataset
-
-
-def read_general(config: Config) -> General:
-    """Read country names and ISO from UNFCCC table."""
-    print("- Reading unfccc country data")
-
-    data_root = config.paths.input
-    filename = "UNFCCC_Parties_Groups_noeu.xlsx"
-
-    # Read and transform countries
-    columns = {"Name": "name", "Country ISO Code": "iso"}
-    countries = (
-        pd.read_excel(
-            data_root / filename,
-            sheet_name="Country groups",
-            usecols=columns.keys(),
-        )
-        .rename(columns=columns)
-        .set_index("name")["iso"]
-        .to_dict()
-    )
-
-    # Extend countries with non-country regions
-    regions = {**countries, **_regions.ADDITIONAL_EU_AND_EARTH}
-
-    return General(countries, regions)
-
-
-def read_ssps_refactor(config, regions):
-    """Read GDP and population data from SSPs."""
-    print("- Reading GDP and population data from SSPs")
-
-    # Define input
-    data_root = config.paths.input
-    filename = "SSPs_v2023.xlsx"
-
-    # Read data
-    df = pd.read_excel(data_root / filename, sheet_name="data")
-
-    # Filter for relevant models
-    df = df[(df.Model.isin(["OECD ENV-Growth 2023", "IIASA-WiC POP 2023"]))].drop(
-        columns=["Model", "Unit"]
-    )
-
-    # Convert year columns into row indexes
-    melted = df.melt(id_vars=["Scenario", "Region", "Variable"], var_name="Time")
-    melted["Time"] = melted["Time"].astype(int)
-
-    # Transform to xarray dataset
-    ds = melted.pivot(
-        index=["Scenario", "Region", "Time"], columns="Variable", values="value"
-    ).to_xarray()
-
-    # Split historical from future scenarios
-    hist = ds.sel(Scenario="Historical Reference", Time=slice(1980, 2020))
-    ds = ds.drop_sel(Scenario="Historical Reference")
-
-    # Substitute historical data into the corresponding years of each scenario
-    historical_expanded = hist.expand_dims(Scenario=ds.Scenario)
-    ds.loc[dict(Time=slice(1980, 2020))] = historical_expanded
-
-    # Rename variable
-    ds = ds.rename_vars({"GDP|PPP": "GDP"})
-
-    # Replace region names with ISO codes
-    region_lookup_table = {**regions, **_regions.ADDITIONAL_REGIONS_SSPS}
-    ds["Region"] = list(map(lambda name: region_lookup_table.get(name, "oeps"), ds.Region.values))
-    ds = ds.sortby(ds.Region)
-
-    return ds
-
-
-# TODO: check why different from refactored version; fix; remove.
-def read_ssps(config, regions):
-    print("- Reading GDP and population data from SSPs")
-
-    # Define input
-    data_root = config.paths.input
-    filename = "SSPs_v2023.xlsx"
-
-    for i in range(6):
-        df_ssp = pd.read_excel(
-            data_root / filename,
-            sheet_name="data",
-        )
-        if i >= 1:
-            df_ssp = df_ssp[
-                (df_ssp.Model.isin(["OECD ENV-Growth 2023"]))
-                & (df_ssp.Scenario == "Historical Reference")
-            ]
-        else:
-            df_ssp = df_ssp[
-                (df_ssp.Model.isin(["OECD ENV-Growth 2023", "IIASA-WiC POP 2023"]))
-                & (df_ssp.Scenario.isin(["SSP1", "SSP2", "SSP3", "SSP4", "SSP5"]))
-            ]
-        region_full = np.array(df_ssp.Region)
-        region_iso = []
-        for r in region_full:
-            iso = regions.get(r, None)
-            if iso is None:
-                # TODO: read from self.regions_iso instead ?!
-                iso = _regions.ADDITIONAL_REGIONS_SSPS.get(r, None)
-            if iso is None:
-                print(r)
-                iso = "oeps"
-            region_iso.append(iso)
-        df_ssp["Region"] = region_iso
-        variable = np.array(df_ssp["Variable"])
-        variable[variable == "GDP|PPP"] = "GDP"
-        df_ssp["Variable"] = variable
-        df_ssp = df_ssp.drop(["Model", "Unit"], axis=1)
-        dummy = df_ssp.melt(
-            id_vars=["Scenario", "Region", "Variable"],
-            var_name="Time",
-            value_name="Value",
-        )
-        dummy["Time"] = np.array(dummy["Time"].astype(int))
-        if i >= 1:
-            dummy["Scenario"] = ["SSP" + str(i)] * len(dummy)
-            xr_hist_gdp_i = xr.Dataset.from_dataframe(
-                dummy.pivot(
-                    index=["Scenario", "Region", "Time"],
-                    columns="Variable",
-                    values="Value",
-                )
-            ).sel(Time=[1980, 1985, 1990, 1995, 2000, 2005, 2010, 2015])
-            xr_ssp = xr.merge([xr_ssp, xr_hist_gdp_i])
-        else:
-            xr_ssp = (
-                xr.Dataset.from_dataframe(
-                    dummy.pivot(
-                        index=["Scenario", "Region", "Time"],
-                        columns="Variable",
-                        values="Value",
-                    )
-                )
-                .reindex({"Time": np.arange(2020, 2101, 5)})
-                .reindex({"Time": np.arange(1980, 2101, 5)})
-            )
-
-        return xr_ssp
-
-
-def read_un_population(config, countries) -> UNPopulation:
-    print("- Reading UN population data and gapminder, processed by OWID (for past population)")
-
-    # Define input
-    data_root = config.paths.input
-    filename = "population_HYDE_UNP_Gapminder.csv"
-
-    df_pop = pd.read_csv(data_root / filename)[["Code", "Year", "Population (historical)"]].rename(
-        {"Code": "Region", "Population (historical)": "Population", "Year": "Time"},
-        axis=1,
-    )
-    reg = np.array(df_pop.Region)
-    reg[reg == "OWID_WRL"] = "EARTH"
-    df_pop.Region = reg
-
-    xr_unp_long = (
-        xr.Dataset.from_dataframe(
-            df_pop[df_pop.Region.isin(list(countries.values()) + ["EARTH"])].set_index(
-                ["Region", "Time"]
-            )
-        )
-        / 1e6
-    )
-    xr_unp = xr_unp_long.sel(Time=np.arange(1850, 2000))
-    return UNPopulation(xr_unp, xr_unp_long)
-
-
-def read_hdi_refactor(config, countries, unpopulation: UNPopulation):
-    print("- Read Human Development Index data")
-
-    # Define input
-    data_root = config.paths.input
-    hdi_file = "HDR21-22_Statistical_Annex_HDI_Table.xlsx"
-
-    df = pd.read_excel(data_root / hdi_file, sheet_name="Rawdata")
-
-    # Convert missing data to NaN
-    df.loc[df.HDI == "..", "HDI"] = np.nan
-
-    # Convert country to region ISO codes
-    def get_iso(x):
-        extended_countries = {**countries, **_regions.ADDITIONAL_REGIONS_HDI}
-        return extended_countries.get(x, "unknown")
-
-    df["Region"] = df.Country.map(get_iso)
-    df = df[~(df.Region == "unknown")]
-
-    # Prepare for conversion to xarray
-    df = df.drop(columns="Country").set_index("Region")
-
-    # Insert NaN countries (TODO: I think we could just skip this??)
-    # fmt: off
-    nan_countries = [
-        "ALA", "ASM", "AIA", "ABW", "BMU", "ANT", "SCG", "BES", "BVT", "IOT", "VGB",
-        "CYM", "CXR", "CCK", "COK", "CUW", "FLK", "FRO", "GUF", "PYF", "ATF", #"GMB",
-        "GIB", "GRL", "GLP", "GUM", "GGY", "HMD", "VAT", "IMN", "JEY", "MAC", "MTQ",
-        "MYT", "MSR", "NCL", "NIU", "NFK", "MNP", "PCN", "PRI", "REU", "BLM", "SHN",
-        "SPM", "SXM", "SGS", "MAF", "SJM", "TKL", "TCA", "UMI", "VIR", "WLF", "ESH",
-    ]
-    # fmt: on
-    df = pd.concat(
-        [
-            df,
-            pd.Series(index=nan_countries, name="HDI").rename_axis("Region"),
-        ]
-    )
-
-    # Convert to xarray
-    hdi = df.to_xarray().dropna("Region")
-
-    # Add hdi_sh
-    pop_2019 = unpopulation.population_long.sel(Time=2019).Population.drop("Time")
-    with xr.set_options(arithmetic_join="outer"):
-        # "outer join" ensures that all regions are kept, even if they are
-        # missing in one of the terms (data will be NaN)
-        hdi_sh = (hdi.HDI / hdi.HDI.sum() * pop_2019).to_dataset(name="HDIsh")
-
-    # TODO: add NaN entries for ISO codes from countries that are not available in HDI data
-    return hdi, hdi_sh
-
-
-def read_hdi(config, countries, unpopulation):
-    print("- Read Human Development Index data")
-
-    # Define input
-    data_root = config.paths.input
-    regions_file = "AR6_regionclasses.xlsx"
-    hdi_file = "HDR21-22_Statistical_Annex_HDI_Table.xlsx"
-
-    df_regions = pd.read_excel(data_root / regions_file)
-    df_regions = df_regions.sort_values(by=["name"])
-    df_regions = df_regions.sort_index()
-
-    df_hdi_raw = pd.read_excel(data_root / hdi_file, sheet_name="Rawdata")
-    hdi_countries_raw = np.array(df_hdi_raw.Country)
-    hdi_values_raw = np.array(df_hdi_raw.HDI).astype(str)
-    hdi_values_raw[hdi_values_raw == ".."] = "nan"
-    hdi_values_raw = hdi_values_raw.astype(float)
-    hdi_av = np.nanmean(hdi_values_raw)
-
-    # fmt: off
-    nan_countries = [
-        "ALA", "ASM", "AIA", "ABW", "BMU", "ANT", "SCG", "BES", "BVT", "IOT", "VGB",
-        "CYM", "CXR", "CCK", "COK", "CUW", "FLK", "FRO", "GUF", "PYF", "ATF", "GMB",
-        "GIB", "GRL", "GLP", "GUM", "GGY", "HMD", "VAT", "IMN", "JEY", "MAC", "MTQ",
-        "MYT", "MSR", "NCL", "NIU", "NFK", "MNP", "PCN", "PRI", "REU", "BLM", "SHN",
-        "SPM", "SXM", "SGS", "MAF", "SJM", "TKL", "TCA", "UMI", "VIR", "WLF", "ESH",
-    ]
-    # fmt: on
-
-    # Construct new hdi object
-    countries_name = list(countries.keys())
-    countries_iso = list(countries.values())
-
-    hdi_values = np.zeros(len(countries_iso)) + np.nan
-    hdi_sh_values = np.zeros(len(countries_iso)) + np.nan
-
-    for r_i, iso in enumerate(countries_iso):
-        name = countries_name[r_i]
-        value = np.where(hdi_countries_raw == name)[0]
-
-        if len(value) > 0:
-            wh_i = value[0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso in nan_countries:
-            hdi_values[r_i] = np.nan
-        elif iso == "USA":
-            value = np.where(hdi_countries_raw == "United States")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "BHS":
-            value = np.where(hdi_countries_raw == "Bahamas")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "GMB":
-            value = np.where(hdi_countries_raw == "Gambia")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "CPV":
-            value = np.where(hdi_countries_raw == "Cabo Verde")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "BOL":
-            value = np.where(hdi_countries_raw == "Bolivia (Plurinational State of)")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "COD":
-            value = np.where(hdi_countries_raw == "Congo (Democratic Republic of the)")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "COG":
-            value = np.where(hdi_countries_raw == "Congo")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "CZE":
-            value = np.where(hdi_countries_raw == "Czechia")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "EGY":
-            value = np.where(hdi_countries_raw == "Egypt")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "HKG":
-            value = np.where(hdi_countries_raw == "Hong Kong, China (SAR)")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "IRN":
-            value = np.where(hdi_countries_raw == "Iran (Islamic Republic of)")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "PRK":
-            value = np.where(hdi_countries_raw == "Korea (Democratic People's Rep. of)")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "KOR":
-            value = np.where(hdi_countries_raw == "Korea (Republic of)")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "KGZ":
-            value = np.where(hdi_countries_raw == "Kyrgyzstan")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "LAO":
-            value = np.where(hdi_countries_raw == "Lao People's Democratic Republic")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "FSM":
-            value = np.where(hdi_countries_raw == "Micronesia (Federated States of)")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "MDA":
-            value = np.where(hdi_countries_raw == "Moldova (Republic of)")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "STP":
-            value = np.where(hdi_countries_raw == "Sao Tome and Principe")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "SVK":
-            value = np.where(hdi_countries_raw == "Slovakia")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "KNA":
-            value = np.where(hdi_countries_raw == "Saint Kitts and Nevis")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "LCA":
-            value = np.where(hdi_countries_raw == "Saint Lucia")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "VCT":
-            value = np.where(hdi_countries_raw == "Saint Vincent and the Grenadines")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "SWZ":
-            value = np.where(hdi_countries_raw == "Eswatini (Kingdom of)")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "TWN":
-            value = np.where(hdi_countries_raw == "China")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "TZA":
-            value = np.where(hdi_countries_raw == "Tanzania (United Republic of)")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "TUR":
-            value = np.where(hdi_countries_raw == "Türkiye")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "VEN":
-            value = np.where(hdi_countries_raw == "Venezuela (Bolivarian Republic of)")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "VNM":
-            value = np.where(hdi_countries_raw == "Viet Nam")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "PSE":
-            value = np.where(hdi_countries_raw == "Palestine, State of")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        elif iso == "YEM":
-            value = np.where(hdi_countries_raw == "Yemen")[0][0]
-            hdi_values[r_i] = hdi_values_raw[wh_i]
-        else:
-            hdi_values[r_i] = np.nan
-
-        try:
-            pop = float(unpopulation.long.sel(Region=iso, Time=2019).Population)
-        except:
-            pop = np.nan
-        hdi_sh_values[r_i] = hdi_values[r_i] * pop
-
-    hdi_sh_values = hdi_sh_values / np.nansum(hdi_sh_values)
-    df_hdi = {}
-
-    df_hdi["Region"] = countries_iso
-    df_hdi["Name"] = countries_name
-    df_hdi["HDI"] = hdi_values
-    df_hdi = pd.DataFrame(df_hdi)
-    df_hdi = df_hdi[["Region", "HDI"]]
-    dfdummy = df_hdi.set_index(["Region"])
-    xr_hdi = xr.Dataset.from_dataframe(dfdummy)
-
-    df_hdi = {}
-    df_hdi["Region"] = countries_iso
-    df_hdi["Name"] = countries_name
-    df_hdi["HDIsh"] = hdi_sh_values
-    df_hdi = pd.DataFrame(df_hdi)
-    df_hdi = df_hdi[["Region", "HDIsh"]]
-    dfdummy = df_hdi.set_index(["Region"])
-    xr_hdish = xr.Dataset.from_dataframe(dfdummy)
-
-    return xr_hdi, xr_hdish
-
-
-def read_historicalemis_jones(config: Config, regions) -> JonesData:
-    # No harmonization with the KEV anymore, but it's also much closer now
-    print("- Reading historical emissions (jones)")
-
-    # Define input
-    # TODO: separate functions for reading each of these files?
-    data_root = config.paths.input
-    guetschow_et_al = "Guetschow_et_al_2024-PRIMAP-hist_v2.5.1_final_no_rounding_27-Feb-2024.nc"
-    emissions_file = "EMISSIONS_ANNUAL_1830-2022.csv"
-    edgar_file = "EDGARv8.0_FT2022_GHG_booklet_2023.xlsx"
-    country_groups_file = "UNFCCC_Parties_Groups_noeu.xlsx"
-
-    # Read data
-    xr_primap2 = xr.open_dataset(data_root / guetschow_et_al)
-    df_nwc = pd.read_csv(data_root / emissions_file)
-    xr_nwc = xr.Dataset.from_dataframe(
-        df_nwc.drop(columns=["CNTR_NAME", "Unit"]).set_index(["ISO3", "Gas", "Component", "Year"])
-    )
-
-    # Rename GLOBAL to EARTH
-    regs = np.array(xr_nwc.ISO3)
-    regs[regs == "GLOBAL"] = "EARTH"
-    xr_nwc["ISO3"] = regs
-
-    # Calculate total(?)
-    xr_nwc_tot = (
-        xr_nwc.sel(Gas="CH[4]") * config.params.gwp_ch4 / 1e3
-        + xr_nwc.sel(Gas="N[2]*O") * config.params.gwp_n2o / 1e3
-        + xr_nwc.sel(Gas="CO[2]") * 1
-    ).drop_vars(["Gas"])
-
-    # Calculate individual contributions
-    # TODO: adding these will yield same total as above and seems cleaner
-    xr_nwc_co2 = xr_nwc.sel(Gas="CO[2]").drop_vars(["Gas"])
-    xr_nwc_ch4 = xr_nwc.sel(Gas="CH[4]").drop_vars(["Gas"]) * config.params.gwp_ch4 / 1e3
-    xr_nwc_n2o = xr_nwc.sel(Gas="N[2]*O").drop_vars(["Gas"]) * config.params.gwp_n2o / 1e3
-
-    # Select relevant data from primap (?)
-    xr_primap_agri = (
-        xr_primap2["KYOTOGHG (AR6GWP100)"]
-        .rename(
-            {
-                "area (ISO3)": "Region",
-                "scenario (PRIMAP-hist)": "scen",
-                "category (IPCC2006_PRIMAP)": "cat",
-            }
-        )
-        .sel(
-            scen="HISTTP",
-            provenance="derived",
-            cat=["M.AG"],
-            source="PRIMAP-hist_v2.5.1_final_nr",
-        )
-        .sum(dim="cat")
-        .drop_vars(["source", "provenance", "scen"])
-    )
-    xr_primap_agri["time"] = np.arange(1750, 2023)
-    xr_primap_agri = xr_primap_agri.rename({"time": "Time"})
-
-    # Same for CO2 from primap (?)
-    xr_primap_agri_co2 = (
-        xr_primap2["CO2"]
-        .rename(
-            {
-                "area (ISO3)": "Region",
-                "scenario (PRIMAP-hist)": "scen",
-                "category (IPCC2006_PRIMAP)": "cat",
-            }
-        )
-        .sel(
-            scen="HISTTP",
-            provenance="derived",
-            cat=["M.AG"],
-            source="PRIMAP-hist_v2.5.1_final_nr",
-        )
-        .sum(dim="cat")
-        .drop_vars(["source", "provenance", "scen"])
-    )
-    xr_primap_agri_co2["time"] = np.arange(1750, 2023)
-    xr_primap_agri_co2 = xr_primap_agri_co2.rename({"time": "Time"})
-
-    # Select historical data from NWC
-    xr_ghghist = (
-        xr_nwc_tot.rename({"ISO3": "Region", "Year": "Time", "Data": "GHG_hist"})
-        .sel(Component="Total")
-        .drop_vars("Component")
-    )
-    xr_co2hist = (
-        xr_nwc_co2.rename({"ISO3": "Region", "Year": "Time", "Data": "CO2_hist"})
-        .sel(Component="Total")
-        .drop_vars("Component")
-    )
-    xr_ch4hist = (
-        xr_nwc_ch4.rename({"ISO3": "Region", "Year": "Time", "Data": "CH4_hist"})
-        .sel(Component="Total")
-        .drop_vars("Component")
-    )
-    xr_n2ohist = (
-        xr_nwc_n2o.rename({"ISO3": "Region", "Year": "Time", "Data": "N2O_hist"})
-        .sel(Component="Total")
-        .drop_vars("Component")
-    )
-
-    # Calculate emissions excluding LULUCF
-    xr_ghgexcl = (
-        xr_nwc_tot.rename({"ISO3": "Region", "Year": "Time"})
-        .sel(Component="Total")
-        .drop_vars("Component")
-        - xr_nwc_tot.rename({"ISO3": "Region", "Year": "Time"})
-        .sel(Component="LULUCF")
-        .drop_vars("Component")
-        + xr_primap_agri / 1e6
-    ).rename({"Data": "GHG_hist_excl"})
-    xr_co2excl = (
-        xr_nwc_co2.rename({"ISO3": "Region", "Year": "Time"})
-        .sel(Component="Total")
-        .drop_vars("Component")
-        - xr_nwc_co2.rename({"ISO3": "Region", "Year": "Time"})
-        .sel(Component="LULUCF")
-        .drop_vars("Component")
-        + xr_primap_agri_co2 / 1e6
-    ).rename({"Data": "CO2_hist_excl"})
-
-    # Combine historical data into single xarray dataset
-    regions_iso = list(regions.values())
-    xr_hist = (
-        xr.merge([xr_ghghist, xr_ghgexcl, xr_co2hist, xr_co2excl, xr_ch4hist, xr_n2ohist]) * 1e3
-    ).reindex({"Region": regions_iso})
-
-    # Store LULUCF (?)
-    xr_ghg_afolu = (
-        xr_nwc_tot.rename({"ISO3": "Region", "Year": "Time"})
-        .sel(Component="LULUCF")
-        .drop_vars("Component")
-    )
-
-    # Change units of agri
-    # TODO: move this line closer to other agri steps?
-    xr_ghg_agri = xr_primap_agri / 1e6
-
-    # Also read EDGAR for purposes of using CR data (note that this is GHG excl LULUCF)
-    # TODO: move to separate function?
-    df_edgar = (
-        pd.read_excel(data_root / edgar_file, sheet_name="GHG_totals_by_country")
-        .drop(["Country"], axis=1)
-        .set_index("EDGAR Country Code")
-    )
-    df_edgar.columns = df_edgar.columns.astype(int)
-
-    # drop second-to-last row
-    df_edgar = df_edgar.drop(df_edgar.index[-2])
-
-    # Rename index column
-    df_edgar.index.name = "Region"
-
-    # Melt time columns into a time index
-    df_edgar = (
-        df_edgar.reset_index()
-        .melt(id_vars="Region", var_name="Time", value_name="Emissions")
-        .set_index(["Region", "Time"])
-    )
-
-    # Convert to xarray
-    xr_edgar = df_edgar.to_xarray()
-
-    xr_primap = xr_primap2.rename(
-        {
-            "area (ISO3)": "Region",
-            "scenario (PRIMAP-hist)": "Scenario",
-            "category (IPCC2006_PRIMAP)": "Category",
-        }
-    ).sel(provenance="derived", source="PRIMAP-hist_v2.5.1_final_nr")
-    xr_primap = xr_primap.assign_coords(time=xr_primap.time.dt.year)
-
-    # Add EU (this is required for the NDC data reading)
-    df = pd.read_excel(data_root / country_groups_file, sheet_name="Country groups")
-    countries_iso = np.array(df["Country ISO Code"])
-    group_eu = countries_iso[np.array(df["EU"]) == 1]
-    xr_hist.GHG_hist.loc[dict(Region="EU")] = xr_hist.GHG_hist.sel(Region=group_eu).sum("Region")
-
-    return JonesData(xr_hist, xr_ghg_afolu, xr_ghg_agri, xr_edgar, xr_primap)
-
-
-def read_ar6(config: Config, xr_hist) -> AR6Data:
-    print("- Read AR6 data")
-
-    # Define input
-    data_root = config.paths.input
-    filename = "AR6_Scenarios_Database_World_v1.1.csv"
-    metadata_file = "AR6_Scenarios_Database_metadata_indicators_v1.1.xlsx"
-    elevate_snapshot = "elevate-internal_snapshot_1739887620.csv"
-
-    df_ar6raw = pd.read_csv(data_root / filename)
-    df_ar6 = df_ar6raw[
-        df_ar6raw.Variable.isin(
-            [
-                "Emissions|CO2",
-                "Emissions|CO2|AFOLU",
-                "Emissions|Kyoto Gases",
-                "Emissions|CO2|Energy and Industrial Processes",
-                "Emissions|CH4",
-                "Emissions|N2O",
-                "Emissions|CO2|AFOLU|Land",
-                "Emissions|CH4|AFOLU|Land",
-                "Emissions|N2O|AFOLU|Land",
-                "Carbon Sequestration|CCS",
-                "Carbon Sequestration|Land Use",
-                "Carbon Sequestration|Direct Air Capture",
-                "Carbon Sequestration|Enhanced Weathering",
-                "Carbon Sequestration|Other",
-                "Carbon Sequestration|Feedstocks",
-                "AR6 climate diagnostics|Exceedance Probability 1.5C|MAGICCv7.5.3",
-                "AR6 climate diagnostics|Exceedance Probability 2.0C|MAGICCv7.5.3",
-                "AR6 climate diagnostics|Exceedance Probability 2.5C|MAGICCv7.5.3",
-                "AR6 climate diagnostics|Exceedance Probability 3.0C|MAGICCv7.5.3",
-                "AR6 climate diagnostics|Surface Temperature (GSAT)|MAGICCv7.5.3|5.0th Percentile",
-                "AR6 climate diagnostics|Surface Temperature (GSAT)|MAGICCv7.5.3|33.0th Percentile",
-                "AR6 climate diagnostics|Surface Temperature (GSAT)|MAGICCv7.5.3|50.0th Percentile",
-                "AR6 climate diagnostics|Surface Temperature (GSAT)|MAGICCv7.5.3|67.0th Percentile",
-                "AR6 climate diagnostics|Surface Temperature (GSAT)|MAGICCv7.5.3|95.0th Percentile",
-            ]
-        )
-    ]
-    df_ar6 = df_ar6.reset_index(drop=True)
-    idx = (
-        df_ar6[(df_ar6.Variable == "Emissions|CH4") & (df_ar6["2100"] > 1e5)]
-    ).index  # Removing erroneous CH4 scenarios
-    df_ar6 = df_ar6[~df_ar6.index.isin(idx)]
-    df_ar6 = df_ar6.reset_index(drop=True)
-
-    df_ar6_meta = pd.read_excel(data_root / metadata_file, sheet_name="meta_Ch3vetted_withclimate")
-    mods = np.array(df_ar6_meta.Model)
-    scens = np.array(df_ar6_meta.Scenario)
-    modscens_meta = np.array([mods[i] + "|" + scens[i] for i in range(len(scens))])
-    df_ar6_meta["ModelScenario"] = modscens_meta
-    df_ar6_meta = df_ar6_meta[["ModelScenario", "Category", "Policy_category"]]
-    mods = np.array(df_ar6.Model)
-    scens = np.array(df_ar6.Scenario)
-    modscens = np.array([mods[i] + "|" + scens[i] for i in range(len(scens))])
-    df_ar6["ModelScenario"] = modscens
-    df_ar6 = df_ar6.drop(["Model", "Scenario", "Region", "Unit"], axis=1)
-    dummy = df_ar6.melt(id_vars=["ModelScenario", "Variable"], var_name="Time", value_name="Value")
-    dummy["Time"] = np.array(dummy["Time"].astype(int))
-    dummy = dummy.set_index(["ModelScenario", "Variable", "Time"])
-    xr_scen2 = xr.Dataset.from_dataframe(dummy)
-    xr_scen2 = xr_scen2.reindex(Time=np.arange(2000, 2101, 10))
-    xr_scen2 = xr_scen2.reindex(Time=np.arange(2000, 2101))
-    xr_ar6_prevet = xr_scen2.interpolate_na(dim="Time", method="linear")
-
-    recent_increment = int(config.params.start_year_analysis // 5 * 5)
-    vetting_nans = np.array(
-        xr_ar6_prevet.ModelScenario[
-            ~np.isnan(xr_ar6_prevet.Value.sel(Time=2100, Variable="Emissions|CO2"))
-        ]
-    )
-    vetting_recentyear = np.array(
-        xr_ar6_prevet.ModelScenario[
-            np.where(
-                np.abs(
-                    xr_ar6_prevet.sel(Time=recent_increment, Variable="Emissions|CO2").Value
-                    - xr_hist.sel(Region="EARTH", Time=recent_increment).CO2_hist
-                )
-                < 1e4
-            )[0]
-        ]
-    )
-    vetting_total = np.intersect1d(vetting_nans, vetting_recentyear)
-    xr_ar6 = xr_ar6_prevet.sel(ModelScenario=vetting_total)
-    ms_immediate = np.array(
-        df_ar6_meta[df_ar6_meta.Policy_category.isin(["P2", "P2a", "P2b", "P2c"])].ModelScenario
-    )
-    ms_delayed = np.array(
-        df_ar6_meta[df_ar6_meta.Policy_category.isin(["P3a", "P3b", "P3c"])].ModelScenario
-    )
-
-    # TODO: shouldn't ch4 also be divided by 1000? That's what was done above in read_jones...
-    xr_ar6_landuse = (
-        xr_ar6.sel(Variable="Emissions|CO2|AFOLU|Land") * 1
-        + xr_ar6.sel(Variable="Emissions|CH4|AFOLU|Land") * config.params.gwp_ch4
-        + xr_ar6.sel(Variable="Emissions|N2O|AFOLU|Land") * config.params.gwp_n2o / 1000
-    )
-    xr_ar6_landuse = xr_ar6_landuse.rename({"Value": "GHG_LULUCF"})
-    xr_ar6_landuse = xr_ar6_landuse.assign(
-        CO2_LULUCF=xr_ar6.sel(Variable="Emissions|CO2|AFOLU|Land").Value
-    )
-
-    # Take averages of GHG excluding land use for the C-categories (useful for the Robiou paper)
-    xr_both = xr.merge([xr_ar6, xr_ar6_landuse])
-    xr_ar6_nozeros = xr_both.where(xr_both > -1e9, np.nan).where(xr_both != 0, np.nan)
-    xr_averages = []
-    for i in range(6):
-        C = [["C1"], ["C1", "C2"], ["C2"], ["C3"], ["C6"], ["C7"]][i]
-        Cname = ["C1", "C1+C2", "C2", "C3", "C6", "C7"][i]
-        C_cat = np.intersect1d(
-            np.array(xr_ar6_nozeros.ModelScenario),
-            np.array(df_ar6_meta[df_ar6_meta.Category.isin(C)].ModelScenario),
-        )
-        xr_averages.append(
-            xr_ar6_nozeros.sel(ModelScenario=C_cat)
-            .mean(dim="ModelScenario")
-            .expand_dims(Category=[Cname])
-        )
-    xr_av = xr.merge(xr_averages)
-    xr_ar6_C = xr.merge(
-        [
-            (xr_av.Value.sel(Variable="Emissions|Kyoto Gases") - xr_av.GHG_LULUCF)
-            .to_dataset(name="GHG_excl_C")
-            .drop_vars("Variable"),
-            (xr_av.Value.sel(Variable="Emissions|CO2") - xr_av.CO2_LULUCF)
-            .to_dataset(name="CO2_excl_C")
-            .drop_vars("Variable"),
-            (
-                xr_av.Value.sel(
-                    Variable=[
-                        "Carbon Sequestration|CCS",
-                        "Carbon Sequestration|Direct Air Capture",
-                    ]
-                ).sum(dim="Variable", skipna=False)
-            ).to_dataset(name="CO2_neg_C"),
-        ]
-    )
-    xr_ar6_C = xr_ar6_C.reindex(Time=np.arange(2000, 2101, 10))
-    xr_ar6_C = xr_ar6_C.reindex(Time=np.arange(2000, 2101))
-
-    # Bunker subtraction
-    # TODO: move to separate function?
-    df_elevate_bunkers = pd.read_csv(data_root / elevate_snapshot)[:-1]
-    mods = np.array(df_elevate_bunkers.Model)
-    scens = np.array(df_elevate_bunkers.Scenario)
-    modscens = np.array([mods[i] + "|" + scens[i] for i in range(len(scens))])
-    df_elevate_bunkers["ModelScenario"] = modscens
-    df_elevate_bunkers = df_elevate_bunkers.drop(["Model", "Scenario", "Region", "Unit"], axis=1)
-    dummy = df_elevate_bunkers.melt(
-        id_vars=["ModelScenario", "Variable"], var_name="Time", value_name="Value"
-    )
-    dummy["Time"] = np.array(dummy["Time"].astype(int))
-    dummy = dummy.set_index(["ModelScenario", "Variable", "Time"])
-    xr_elevate_bunkers = xr.Dataset.from_dataframe(dummy)
-    xr_elevate_bunkers = xr_elevate_bunkers.reindex({"Time": np.arange(2010, 2101, 10)})
-
-    modscens = np.array(xr_elevate_bunkers.ModelScenario)
-    categories = []
-    for ms in modscens:
-        if (
-            xr_elevate_bunkers.sel(
-                ModelScenario=ms,
-                Variable="AR6 climate diagnostics|Surface Temperature (GSAT)|MAGICCv7.5.3|50.0th Percentile",
-            )
-            .max()
-            .Value
-            < 1.5
-        ):
-            categories.append("C1")
-        elif (
-            xr_elevate_bunkers.sel(
-                ModelScenario=ms,
-                Variable="AR6 climate diagnostics|Surface Temperature (GSAT)|MAGICCv7.5.3|50.0th Percentile",
-            )
-            .max()
-            .Value
-            < 1.7
-        ):
-            categories.append("C2")
-        elif (
-            xr_elevate_bunkers.sel(
-                ModelScenario=ms,
-                Variable="AR6 climate diagnostics|Surface Temperature (GSAT)|MAGICCv7.5.3|50.0th Percentile",
-            )
-            .max()
-            .Value
-            < 2.0
-        ):
-            categories.append("C3")
-        elif (
-            xr_elevate_bunkers.sel(
-                ModelScenario=ms,
-                Variable="AR6 climate diagnostics|Surface Temperature (GSAT)|MAGICCv7.5.3|50.0th Percentile",
-            )
-            .max()
-            .Value
-            < 3.0
-        ):
-            categories.append("C6")
-        elif (
-            xr_elevate_bunkers.sel(
-                ModelScenario=ms,
-                Variable="AR6 climate diagnostics|Surface Temperature (GSAT)|MAGICCv7.5.3|50.0th Percentile",
-            )
-            .max()
-            .Value
-            < 4.0
-        ):
-            categories.append("C7")
-        elif (
-            xr_elevate_bunkers.sel(
-                ModelScenario=ms,
-                Variable="AR6 climate diagnostics|Surface Temperature (GSAT)|MAGICCv7.5.3|50.0th Percentile",
-            )
-            .max()
-            .Value
-            > 4.0
-        ):
-            categories.append("C8")
-        else:
-            categories.append(
-                "C9"
-            )  # Fictuous category to show that for these mds, there is no temperature assessment
-    categories = np.array(categories)
-
-    xrs = []
-    for temp in [1.7, 2.0, 3.0]:
-        if temp == 1.7:
-            cat = "C2"
-        elif temp == 2.0:
-            cat = "C3"
-        elif temp == 3.0:
-            cat = "C6"
-        xrs.append(
-            (
-                xr_elevate_bunkers.sel(
-                    Variable="Emissions|CO2|Energy|Demand|Bunkers",
-                    ModelScenario=modscens[categories == cat],
-                ).median(dim="ModelScenario")
-            ).expand_dims(Temperature=[temp])
-        )
-    xr_all = xr.concat(xrs, dim="Temperature")
-    xr_all = xr_all.reindex(Temperature=[1.5, 1.6, 1.7, 2.0, 3.0, 4.0, 4.5])
-
-    # Extrapolation
-    vals = xr_all.loc[dict(Temperature=3.0)] - xr_all.loc[dict(Temperature=1.7)]
-    vals = vals.where(vals > 0, 0)
-    xr_all.loc[dict(Temperature=1.5)] = xr_all.loc[dict(Temperature=1.7)] - vals * (0.2 / 1.3)
-    xr_all.loc[dict(Temperature=1.6)] = xr_all.loc[dict(Temperature=1.7)] - vals * (0.1 / 1.3)
-    xr_all.loc[dict(Temperature=4.0)] = xr_all.loc[dict(Temperature=3.0)] + vals * (1.0 / 1.3)
-    xr_all.loc[dict(Temperature=4.5)] = xr_all.loc[dict(Temperature=3.0)] + vals * (1.5 / 1.3)
-
-    xr_all = (
-        xr_all.rename({"Temperature": "Category"})
-        .drop_vars("Variable")
-        .rename({"Value": "CO2_bunkers_C"})
-    )
-
-    # Rename ticks of temperature
-    xr_all = xr_all.assign_coords(Category=["C1", "C1+C2", "C2", "C3", "C6", "C7", "C8"])
-    xr_ar6_C_bunkers = xr_all
-
-    return AR6Data(
-        xr_ar6_prevet,
-        xr_ar6,
-        ms_immediate,
-        ms_delayed,
-        xr_ar6_landuse,
-        xr_ar6_C,
-        xr_ar6_C_bunkers,
-    )
 
 
 def nonco2variation(config: Config):
@@ -1655,346 +762,6 @@ def determine_global_co2_trajectories(
         all_projected_gases,
     )
 
-
-def read_baseline(
-    config: Config,
-    countries,  # TODO: pass in region instead??
-    xr_hist,
-):
-    print("- Reading baseline emissions")
-
-    data_root = config.paths.input
-    start_year = config.params.start_year_analysis
-    countries_iso = list(countries.values())
-
-    xr_bases = []
-    for i in range(3):
-        # In the up-to-date baselines, only SSP1, 2 and 3 are included. Will be updated at some point.
-        df_base = pd.read_excel(data_root / f"SSP{i + 1}.xlsx", sheet_name="Sheet1")
-        df_base = df_base[df_base["Unnamed: 1"] == "Emissions|CO2|Energy"]
-        df_base = df_base.drop(["Unnamed: 1"], axis=1)
-        df_base = df_base.rename(columns={"COUNTRY": "Region"})
-        df_base["Scenario"] = ["SSP" + str(i + 1)] * len(df_base)
-
-        # Melt time index
-        df_base = df_base.melt(
-            id_vars=["Region", "Scenario"], var_name="Time", value_name="CO2_base_excl"
-        )
-        df_base["Time"] = np.array(df_base["Time"].astype(int))
-
-        # Convert to xarray
-        dummy = df_base.set_index(["Region", "Scenario", "Time"])
-        dummy = dummy.astype(float)
-        xr_bases.append(xr.Dataset.from_dataframe(dummy))
-
-    xr_base = xr.merge(xr_bases).reindex({"Region": countries_iso})
-
-    # Assign 2020 values in Time index
-    xr_base = xr_base.reindex(Time=np.arange(start_year, 2101))
-    for year in np.arange(start_year, 2021):
-        xr_base.CO2_base_excl.loc[dict(Time=year, Region=countries_iso)] = xr_hist.sel(
-            Time=year, Region=countries_iso
-        ).CO2_hist_excl
-
-    # TODO: might be useful to create a helper function like so:
-    def mask_outside(data, lower=-1e9, upper=1e9):
-        cond1 = data > lower
-        cond2 = data < upper
-        return data.where(cond1 & cond2)
-
-    # Harmonize emissions from historical values to baseline emissions
-    diffrac = xr_hist.CO2_hist_excl.sel(Time=start_year) / xr_base.CO2_base_excl.sel(
-        Time=start_year
-    )
-    diffrac = diffrac.where(diffrac < 1e9)
-    diffrac = diffrac.where(diffrac > -1e9)
-    xr_base = xr_base.assign(CO2_base_excl=xr_base.CO2_base_excl * diffrac)
-
-    # Using a fraction, get other emissions variables
-    fraction_startyear_co2_incl = (
-        xr_hist.sel(Time=start_year).CO2_hist / xr_hist.sel(Time=start_year).CO2_hist_excl
-    )
-    fraction_startyear_co2_incl = fraction_startyear_co2_incl.where(
-        fraction_startyear_co2_incl < 1e9
-    )
-    fraction_startyear_co2_incl = fraction_startyear_co2_incl.where(
-        fraction_startyear_co2_incl > -1e9
-    )
-    xr_base = xr_base.assign(CO2_base_incl=xr_base.CO2_base_excl * fraction_startyear_co2_incl)
-
-    fraction_startyear_ghg_excl = (
-        xr_hist.sel(Time=start_year).GHG_hist_excl / xr_hist.sel(Time=start_year).CO2_hist_excl
-    )
-    fraction_startyear_ghg_excl = fraction_startyear_ghg_excl.where(
-        fraction_startyear_ghg_excl < 1e9
-    )
-    fraction_startyear_ghg_excl = fraction_startyear_ghg_excl.where(
-        fraction_startyear_ghg_excl > -1e9
-    )
-    xr_base = xr_base.assign(GHG_base_excl=xr_base.CO2_base_excl * fraction_startyear_ghg_excl)
-
-    fraction_startyear_ghg_incl = (
-        xr_hist.sel(Time=start_year).GHG_hist / xr_hist.sel(Time=start_year).CO2_hist_excl
-    )
-    fraction_startyear_ghg_incl = fraction_startyear_ghg_incl.where(
-        fraction_startyear_ghg_incl < 1e9
-    )
-    fraction_startyear_ghg_incl = fraction_startyear_ghg_incl.where(
-        fraction_startyear_ghg_incl > -1e9
-    )
-    xr_base = xr_base.assign(GHG_base_incl=xr_base.CO2_base_excl * fraction_startyear_ghg_incl)
-
-    # Assign 2020 values in Time index
-    xr_base = xr_base.reindex(Time=np.arange(start_year, 2101))
-    for year in np.arange(start_year, 2021):
-        xr_base.GHG_base_excl.loc[dict(Time=year, Region=countries_iso)] = xr_hist.sel(
-            Time=year, Region=countries_iso
-        ).GHG_hist_excl
-        xr_base.CO2_base_incl.loc[dict(Time=year, Region=countries_iso)] = xr_hist.sel(
-            Time=year, Region=countries_iso
-        ).CO2_hist
-        xr_base.GHG_base_incl.loc[dict(Time=year, Region=countries_iso)] = xr_hist.sel(
-            Time=year, Region=countries_iso
-        ).GHG_hist
-
-    # Harmonize global baseline emissions with sum of all countries (this is important for consistency of AP, etc.)
-    base_onlyc = xr_base.reindex(Region=countries_iso)
-    base_w = base_onlyc.sum(dim="Region").expand_dims({"Region": ["EARTH"]})
-    xr_base = xr.merge([base_w, base_onlyc])
-
-    return xr_base
-
-
-def read_ndc_climateresource(config: Config, countries):
-    print("- Reading NDC data from Climate resource")
-
-    countries_iso = list(countries.values())
-    version_ndcs = config.params.version_ndcs
-    data_root = config.paths.input / f"ClimateResource_{version_ndcs}"
-
-    ghg_data = np.zeros(shape=(len(countries_iso) + 1, 3, 2, 2, len(np.arange(2010, 2051))))
-    for cty_i, cty in enumerate(countries_iso):
-        for cond_i, cond in enumerate(["conditional", "range", "unconditional"]):
-            for hot_i, hot in enumerate(["include", "exclude"]):
-                for amb_i, amb in enumerate(["low", "high"]):
-                    filename = f"{cty.lower()}_ndc_{version_ndcs}_CR_{cond}_{hot}.json"
-                    path = data_root / cond / hot / filename
-                    try:
-                        with open(path) as file:
-                            json_data = json.load(file)
-                        country_name = json_data["results"]["country"]["name"]
-                        series_items = json_data["results"]["series"]
-                        for item in series_items:
-                            columns = item["columns"]
-                            if (
-                                columns["variable"] == "Emissions|Total GHG excl. LULUCF"
-                                and columns["category"] == "Updated NDC"
-                                and columns["ambition"] == amb
-                            ):
-                                data = item["data"]
-                                time_values = [int(year) for year in data.keys()]
-                                ghg_values = np.array(list(item["data"].values()))
-                                ghg_values[ghg_values == "None"] = np.nan
-                                ghg_values = ghg_values.astype(float)
-                                ghg_values = ghg_values[np.array(time_values) >= 2010]
-                                ghg_data[cty_i, cond_i, hot_i, amb_i] = ghg_values
-                                # series.append([country_iso.upper(), country_name, "Emissions|Total GHG excl. LULUCF", conditionality, hot_air, ambition] + list(ghg_values))
-                    except:
-                        continue
-
-    # Now also for EU
-    for cond_i, cond in enumerate(["conditional", "range", "unconditional"]):
-        for hot_i, hot in enumerate(["include", "exclude"]):
-            for amb_i, amb in enumerate(["low", "high"]):
-                filename = f"groupeu27_ndc_{version_ndcs}_CR_{cond}_{hot}.json"
-                path = data_root / cond / hot / "regions" / filename
-                try:
-                    with open(path) as file:
-                        json_data = json.load(file)
-                    country_name = json_data["results"]["country"]["name"]
-                    series_items = json_data["results"]["series"]
-                    for item in series_items:
-                        columns = item["columns"]
-                        if (
-                            columns["variable"] == "Emissions|Total GHG excl. LULUCF"
-                            and columns["category"] == "Updated NDC"
-                            and columns["ambition"] == amb
-                        ):
-                            data = item["data"]
-                            time_values = [int(year) for year in data.keys()]
-                            ghg_values = np.array(list(item["data"].values()))
-                            ghg_values[ghg_values == "None"] = np.nan
-                            ghg_values = ghg_values.astype(float)
-                            ghg_values = ghg_values[np.array(time_values) >= 2010]
-                            ghg_data[cty_i + 1, cond_i, hot_i, amb_i] = ghg_values
-                            # series.append([country_iso.upper(), country_name, "Emissions|Total GHG excl. LULUCF", conditionality, hot_air, ambition] + list(ghg_values))
-                except:
-                    continue
-
-    coords = {
-        "Region": list(countries_iso) + ["EU"],
-        "Conditionality": ["conditional", "range", "unconditional"],
-        "Hot_air": ["include", "exclude"],
-        "Ambition": ["min", "max"],
-        "Time": np.array(time_values)[np.array(time_values) >= 2010],
-    }
-    data_vars = {
-        "GHG_ndc_excl_CR": (
-            ["Region", "Conditionality", "Hot_air", "Ambition", "Time"],
-            ghg_data,
-        ),
-    }
-    xr_ndc = xr.Dataset(data_vars, coords=coords)
-    xr_ndc_CR = xr_ndc.sel(Time=2030)
-
-    return xr_ndc_CR
-
-
-def read_ndc(config: Config, countries, xr_hist):
-    print("- Reading NDC data")
-
-    data_root = config.paths.input
-    filename = "Infographics PBL NDC Tool 4Oct2024_for CarbonBudgetExplorer.xlsx"
-
-    # TODO: use package or better dict mapping method instead
-    countries_name = np.array(list(countries.keys()))
-    countries_iso = np.array(list(countries.values()))
-
-    df_ndc_raw = pd.read_excel(
-        data_root / filename, sheet_name="Reduction All_GHG_incl", header=[0, 1]
-    )
-    regs = df_ndc_raw["(Mt CO2 equivalent)"]["Country name"]
-    regs_iso = []
-
-    for r in regs:
-        wh = np.where(countries_name == r)[0]
-        if len(wh) == 0:
-            if r == "United States":
-                regs_iso.append("USA")
-            elif r == "EU27":
-                regs_iso.append("EU")
-            elif r == "Turkey":
-                regs_iso.append("TUR")
-            else:
-                regs_iso.append(np.nan)
-        else:
-            regs_iso.append(countries_iso[wh[0]])
-    regs_iso = np.array(regs_iso)
-    df_ndc_raw["ISO"] = regs_iso
-
-    df_regs = []
-    df_amb = []
-    df_con = []
-    df_emis = []
-    df_lulucf = []
-    df_red = []
-    df_abs = []
-    df_inv = []
-    histemis = xr_hist.GHG_hist.sel(Time=2015)
-    for r in list(countries_iso) + ["EU"]:
-        histemis_r = float(histemis.sel(Region=r))
-        df_ndc_raw_sub = df_ndc_raw[df_ndc_raw["ISO"] == r]
-        if len(df_ndc_raw_sub) > 0:
-            val_2015 = float(df_ndc_raw_sub["(Mt CO2 equivalent)"][2015])
-            for lulucf in ["incl"]:  # Maybe add excl later?
-                for emis_i, emis in enumerate(["NDC"]):  # , 'CP']):
-                    key = ["2030 NDCs", "Domestic actions 2030"][emis_i]
-                    for cond_i, cond in enumerate(["unconditional", "conditional"]):
-                        condkey = ["Unconditional NDCs", "Conditional NDCs"][cond_i]
-                        for ambition_i, ambition in enumerate(["min", "max"]):
-                            add = ["", ".1"][ambition_i]
-                            val = float(df_ndc_raw_sub[key][condkey + add])
-                            red = 1 - val / val_2015
-                            abs_jones = histemis_r * (1 - red)
-                            df_regs.append(r)
-                            df_amb.append(ambition)
-                            df_con.append(cond)
-                            df_emis.append(emis)
-                            df_lulucf.append(lulucf)
-                            df_red.append(red)
-                            df_abs.append(abs_jones)
-                            df_inv.append(val)
-
-    dict_ndc = {
-        "Region": df_regs,
-        "Ambition": df_amb,
-        "Conditionality": df_con,
-        "GHG_ndc_red": df_red,
-        "GHG_ndc": df_abs,
-        "GHG_ndc_inv": df_inv,
-    }
-    df_ndc = pd.DataFrame(dict_ndc)
-    xr_ndc = xr.Dataset.from_dataframe(df_ndc.set_index(["Region", "Ambition", "Conditionality"]))
-
-    # Now for GHG excluding LULUCF
-    df_ndc_raw = pd.read_excel(
-        data_root / filename, sheet_name="Reduction All_GHG_excl", header=[0, 1]
-    )
-    regs = df_ndc_raw["(Mt CO2 equivalent)"]["Country name"]
-    regs_iso = []
-    for r in regs:
-        wh = np.where(countries_name == r)[0]
-        if len(wh) == 0:
-            if r == "United States":
-                regs_iso.append("USA")
-            elif r == "EU27":
-                regs_iso.append("EU")
-            elif r == "Turkey":
-                regs_iso.append("TUR")
-            else:
-                regs_iso.append(np.nan)
-        else:
-            regs_iso.append(countries_iso[wh[0]])
-    regs_iso = np.array(regs_iso)
-    df_ndc_raw["ISO"] = regs_iso
-
-    df_regs = []
-    df_amb = []
-    df_con = []
-    df_emis = []
-    df_lulucf = []
-    df_red = []
-    df_abs = []
-    df_inv = []
-    histemis = xr_hist.GHG_hist.sel(Time=2015)
-    for r in list(countries_iso) + ["EU"]:
-        histemis_r = float(histemis.sel(Region=r))
-        df_ndc_raw_sub = df_ndc_raw[df_ndc_raw["ISO"] == r]
-        if len(df_ndc_raw_sub) > 0:
-            val_2015 = float(df_ndc_raw_sub["(Mt CO2 equivalent)"][2015])
-            for lulucf in ["incl"]:  # Maybe add excl later?
-                for emis_i, emis in enumerate(["NDC"]):  # , 'CP']):
-                    key = ["2030 NDCs", "Domestic actions 2030"][emis_i]
-                    for cond_i, cond in enumerate(["unconditional", "conditional"]):
-                        condkey = ["Unconditional NDCs", "Conditional NDCs"][cond_i]
-                        for ambition_i, ambition in enumerate(["min", "max"]):
-                            add = ["", ".1"][ambition_i]
-                            val = float(df_ndc_raw_sub[key][condkey + add])
-                            red = 1 - val / val_2015
-                            abs_jones = histemis_r * (1 - red)
-                            df_regs.append(r)
-                            df_amb.append(ambition)
-                            df_con.append(cond)
-                            df_emis.append(emis)
-                            df_lulucf.append(lulucf)
-                            df_red.append(red)
-                            df_abs.append(abs_jones)
-                            df_inv.append(val)
-
-    dict_ndc = {
-        "Region": df_regs,
-        "Ambition": df_amb,
-        "Conditionality": df_con,
-        "GHG_ndc_excl_red": df_red,
-        "GHG_ndc_excl": df_abs,
-        "GHG_ndc_excl_inv": df_inv,
-    }
-    df_ndc = pd.DataFrame(dict_ndc)
-    xr_ndc_excl = xr.Dataset.from_dataframe(
-        df_ndc.set_index(["Region", "Ambition", "Conditionality"])
-    )
-
-    return xr_ndc, xr_ndc_excl
 
 def merge_xr(
     xr_ssp,
